… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/BorrowedFields.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/BorrowedFieldsReturning.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/BorrowedFieldsWithBounds.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatBoolArray.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatBoolView.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatCharArray.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatCharView.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatF32Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatF32View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatF64Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatF64View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI16Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI16View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI32Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI32View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI64Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI64View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI8Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatI8View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatIsizeArray.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatIsizeView.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatString16Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatString16View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatStringArray.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatStringView.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatStrings16Array.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatStrings16View.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatStringsArray.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatStringsView.java
… feature_tests/java/somelib/src/main/java/dev/diplomattest/somelib/ntv/DiplomatU16Array.java