@jcrist1 2024-08-14
Okay @Manishearth I think I'm ready to open this up for review. There are still some issues that would be nice to have, but maybe aren't dealbreakers
* I think this could do with a lot more testing, but I'd kind of like to solicit some ideas. Maybe if someone has a concrete use case they can propose some additional test cases (@brainstorm you said you were interested in the Panama backend do you have any examples you'd like to specifically see?). 
* I'm also a little embarrassed to admit that I can't figure out how to run the tests from the command line (as a reminder my background is in scala, so I'm still pretty new to plain java) so that's still open. 
* Similar to the kotlin backend, I'm not entirely convinced I have the borrows correct. 
* Finally in order to get this working on CI the CI instance would need jextract. I'm open openning up a PR somewhere to add it, but don't know where to start as I've not worked with Github Actions before.

Also it never feels good opening a PR that looks like this: 
<img width="136" alt="Screenshot 2024-08-14 at 21 23 33" src="https://github.com/user-attachments/assets/96d55d71-a9c2-4b03-b3d1-b63fe3982c93">
😅