@jcrist1 2024-08-15
This will probably sit for a little bit again while I'll be looking into the kotlin ci setup, as that will be a good intro to github actions so I can set something up later for this.