@jcrist1 2024-09-01
I added ci tasks for generating the java code and testing the java code. However I'm having some issues with the test task it doesn't actually do this https://github.com/rust-diplomat/diplomat/actions/runs/10654429057/job/29530731634?pr=499#step:5:67
```
-------------------------------------------------------
 T E S T S
-------------------------------------------------------

Results :

Tests run: 0, Failures: 0, Errors: 0, Skipped: 0
```
When I run `cargo make test-java` locally it actually runs tests.
```
[INFO] -------------------------------------------------------
[INFO]  T E S T S
[INFO] -------------------------------------------------------
[INFO] Running dev.diplomattest.somelib.Float64VecTest
WARNING: A restricted method in java.lang.foreign.SymbolLookup has been called
WARNING: java.lang.foreign.SymbolLookup::libraryLookup has been called by dev.diplomattest.somelib.ntv.somelib_h in an unnamed module
WARNING: Use --enable-native-access=ALL-UNNAMED to avoid a warning for callers in this module
WARNING: Restricted methods will be blocked in a future release unless native access is enabled

[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.157 s -- in dev.diplomattest.somelib.Float64VecTest
[INFO] Running dev.diplomattest.somelib.OpaqueTest
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.017 s -- in dev.diplomattest.somelib.OpaqueTest
[INFO] Running dev.diplomattest.somelib.MyStringTest
[INFO] Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.003 s -- in dev.diplomattest.somelib.MyStringTest
[INFO]
[INFO] Results:
[INFO]
[INFO] Tests run: 3, Failures: 0, Errors: 0, Skipped: 0
[INFO]
```
So I need to to investigate a bit more, but am honestly a bit frustrated with Java right now, and if anyone who knows something about maven and java can help I'd definitely appreciate it. 
`</screaming into void>`