@jcrist1 2024-09-13
Okay @Manishearth I think this could do with another review. I substantially redid the generic native <-> java conversion coded. Also added the ci steps. I'm still missing a ci step that actually runs the tests, but I'd like to wrap this up, and focus on smaller pieces at a time.