@Manishearth 2024-09-19
Alright, so my biggest concern with this is making the JDK a part of the `gen` step. "Run tests for a different backend" isn't a common task for Diplomat development, but "regen test code for all backends" is (since it's common enough to add stuff to the `feature_tests` crate), and this means all Diplomat devs need to set up JDK, something that can often be very finicky.

I would _like_ for the Java backend to be a part of this repo, however. We have a couple routes in front of us, and I'd like to know what you think:

 1. Diplomat devs get used to setting up JDK, potentially with us merging PRs that temporarily break the java gen task (we can keep it a separate CI job) with the understanding that one of us with a JDK setup will fix it. I can commit to doing so, though I'd rather not.
      - Downside: This is going to be really annoying for other Diplomat devs. I'm open to this but I'd rather not go down this route.
 3. This backend is added in-tree, however we do not check in generated code for it, and CI runs `gen-java` before running tests.
     - Downside: we lose out on Java diffs
4. This backend is added in-tree, and we check in regular generated code but not the jextracted code, and CI runs jextract (or perhaps some command like `diplomat-tool --backend jextractwrapper`)
   - Downside: Kind of weird, we still lose out on diffs but not the important ones
5. This backend is maintained as `rust-diplomat/java` as a separate crate, and has full CI there.
   - Downside: Diplomat changes to the HIR don't automatically get nicely propagated there, it's possible for the backends to diverge more in features. Right now even if I don't maintain the Kotlin backend whenever I change something it is relatively easy for me to ensure that Kotlin is kept up to date most of the time.
 6. We manually generate bindings using JEP 454 APIs rather than using `jextract`
   - Downside: this is a bunch of extra work


The last one is my preferred end state, honestly: when we were considering Panama we largely intended for the JEP 454 `java.lang.foreign` (etc) APIs to be used, but not `jextract`. I think `jextract` is a great way to bootstrap this project, though.


A hybrid route I find somewhat appealing is to take option 4 or 5 above ("don't check in the full bindings"), and over time move to a diplomat that doesn't need jextract. A particularly cool option would be for diplomat to generate an `ntv/` folder that is API compatible with jextract, so we can run tests swapping out the diplomat-generated ntv with the jextracted ntv.


@jcrist1 what do you think? Ultimately you're the one doing the work here, so I don't want to dictate anything, but I want to try and resolve the tradeoff with the impact on other backends, and try to understand your own preferences here.